- [x] implement giving coins - [x] implement cache for placeholder, get player current số tiền đã nạp / also async cache thing - [x] implementation database to save payment history

+++ simppay-paper/src/main/java/org/simpmc/simppay/handler/banking/data/BankingData.java
@@ -12,7 +12,7 @@ public class BankingData {
     public UUID playerUUID;
     public String desc;
     public String bin; // represents the bank name
-    public long amount;
+    public double amount;
     public String accountNumber;
     public String url;
